Damar

See merge request yaufaniadam/simkatmawamaret!11

+++ application/modules/admin/controllers/Periode.php
@@ -207,14 +207,19 @@ class Periode extends Admin_Controller
 		redirect(base_url('admin/periode/bulan/' . $id_periode));
 	}
 
-	
-	public function hapus($id, $id_periode, $id_pengajuan)
+
+	public function hapus()
 	{
-		$this->db->delete('Tr_Penerbitan_Pengajuan', array('id_penerbitan_pengajuan' => $id));
-		$this->session->set_flashdata('msg', 'Data berhasil dihapus!');
+		if ($this->input->post('command') == "DELETE") {
+			$id_penerbitan_pengajuan = $this->input->post('id_penerbitan_pengajuan');
+			$id_pengajuan = $this->input->post("id_pengajuan");
+			$id_periode = $this->input->post("id_periode");
 
-		$this->db->delete('Tr_Pengajuan_Status', array('pengajuan_id' => $id_pengajuan, 'status_id'=> 9 ));
-		redirect(base_url('admin/periode/bulan/' . $id_periode));
+			$this->db->delete('Tr_Penerbitan_Pengajuan', array('id_penerbitan_pengajuan' => $id_penerbitan_pengajuan));
+			$this->session->set_flashdata('msg', 'Data berhasil dihapus!');
+
+			$this->db->delete('Tr_Pengajuan_Status', array('pengajuan_id' => $id_pengajuan, 'status_id' => 9));
+			redirect(base_url('admin/periode/bulan/' . $id_periode));
+		}
 	}
-
 }
